doHover: max width/height for float windows
Previously some language servers would emit very long lines in their
documentation, for example, clangd. There was no cap applied to the
width of the resulting window, which would make it very hard to use. I
have, while I was at it, also allowed forcing the height to be limited.